Update handicap system to support decimal values - Change handicap fields from Int to Float in database schema - Add step=0.1 to all handicap number inputs - Replace parseInt with parseFloat for handicap parsing - Update placeholders to show decimal examples (e.g., 17.2) - Support handicaps like 17.2, 21.2 etc across all forms

+++ app/api/admin/members/[id]/route.ts
@@ -47,7 +47,7 @@ export async function PATCH(
         name,
         email,
         phone: phone || null,
-        handicap: parseInt(handicap)
+        handicap: parseFloat(handicap)
       }
     })
